requirements.txt
pandas>=3.0.0
requests>=2.32.5
beautifulsoup4>=4.14.3
python-dotenv>=1.2.1
playwright>=1.58.0